+++ dev-test/codegen.ts
@@ -232,6 +232,17 @@ const config: CodegenConfig = {
         mergeFragmentTypes: true,
       },
     },
+    './dev-test/subpath-import/result.d.ts': {
+      schema: './dev-test/subpath-import/schema.graphql',
+      plugins: ['typescript', 'typescript-resolvers'],
+      config: {
+        contextType: '\\#test-null-value/context#TestContext',
+        fieldContextTypes: ['mutation.createUser#\\#test/root#FiedContextType'],
+        enumValues: {
+          RoleStatus: '\\#changeName/server/drizzle/schema#RoleStatus',
+        },
+      },
+    },
   },
 };