+++ website/tailwind.config.cjs
@@ -1,1 +1,0 @@
-module.exports = require('@theguild/tailwind-config');